chore(package.json): add 'local' script to run local.ts file in playground folder feat(playground): add local.ts file for testing deepl-node translation functionality feat(playground): add nested.json and simple.json files for translation testing
