+++ playground/nested.json
@@ -1,0 +1,8 @@
+{
+    "landing": {
+        "welcome": "Hello, welcome!"
+    },
+    "profile": {
+        "intro": "What is up for snacks later?"
+    }
+}
