+++ playground/simple.json
@@ -1,0 +1,4 @@
+{
+    "welcome": "Hello, welcome!",
+    "intro": "What is up for snacks later?"
+}
